[Gold V] Title: 옥상 정원 꾸미기, Time: 232 ms, Memory: 24160 KB -BaekjoonHub

+++ 백준/Gold/6198. 옥상 정원 꾸미기/옥상 정원 꾸미기.java
@@ -5,29 +5,27 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
-        ArrayDeque<long[]> stack = new ArrayDeque<>();
-        long[] input = new long[n];
+        ArrayDeque<int[]> stack = new ArrayDeque<>();
+        int[] input = new int[n];
 
         for (int i = 0; i < n; i++) {
-            input[i] = Long.parseLong(br.readLine());
+            input[i] = Integer.parseInt(br.readLine());
         }
 
         long result = 0;
 
         for (int i = n - 1; i > -1; i--) {
-            long curHeight = input[i];
+            int curHeight = input[i];
             int canSee = 0;
 
             while (!stack.isEmpty() && (stack.peek()[1] < curHeight)) {
 
-                    long[] removedBuilding = stack.pop();
+                    int[] removedBuilding = stack.pop();
                     canSee += (removedBuilding[0] + 1);
-
-
-
+                    
             }
 
-            stack.push(new long[]{canSee, curHeight});
+            stack.push(new int[]{canSee, curHeight});
             result += canSee;
         }
